pyproject.toml
[tool.poetry]
name = "vectorvision"
version = "0.1.2"
description = "CLI tool for vectorizing raster graphics, written on the basis of the potrace algorithm."
authors = ["Wojciech Łapacz", "Karol Ziarek", "Kajetan Rożej"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alabaster = "0.7.16"
Babel = "2.14.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
contourpy = "1.2.1"
cycler = "0.12.1"
exceptiongroup = "1.2.0"
fonttools = "4.51.0"
idna = "3.6"
imagesize = "1.4.1"
iniconfig = "2.0.0"
Jinja2 = "3.1.3"
kiwisolver = "1.4.5"
MarkupSafe = "2.1.5"
matplotlib = "3.8.4"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
opencv-python = "4.9.0.80"
packaging = "24.0"
pathspec = "0.12.1"
pillow = "10.3.0"
platformdirs = "4.2.0"
pluggy="1.4.0"
Pygments = "2.17.2"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
requests = "2.31.0"
shapely = "2.0.4"
six = "1.16.0"
snowballstemmer = "2.2.0"
tomli = "2.0.1"
typing_extensions = "4.11.0"
urllib3 = "2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vectorvision = "vectorvision.main:main"
